intelligent-bank-risk-control-system/ai_services/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
requests==2.32.3
rapidocr-onnxruntime==1.2.3
Pillow==11.0.0
numpy==2.1.3
python-multipart==0.0.27
scikit-learn==1.5.2
langchain>=0.3.0,<0.4
langchain-community>=0.3.0,<0.4
langchain-core>=0.3.0,<0.4
langchain-ollama>=0.2.0,<0.4
langchain-text-splitters>=0.3.0,<0.4
chromadb>=0.5.0,<1.0
pypdf>=5.0.0,<6
python-docx>=1.1.0,<2
